plugins: odp: [WA] Force inline mode of ODP timer
This is a workaround for following VPP error:
  "Program terminated with signal SIGUSR1, User defined signal 1."
It changes mode of timers used by ODP schedule mode.

By default inline_timers is false and timer_res_init (odp_timer.c) is
called to adjust timer resolution. This function uses SIGUSR1 during
the process which collides with VPP as it uses the same signal for
communication with clients. This patch causes inline_timers to be set
and disables adjustment of timer resolution.

Change-Id: I52a25870c0a01a1688de32b9376f91b7bd720e4d
Signed-off-by: Michal Mazur <mkm@semihalf.com>