odp: Improve readme file, enable build with v1.15
Some features like IPSec and support for i40e driver were affected by
new release of odp-dpdk. I40e fails sometimes to reset NIC and it is
impossible to use DPDK crypto pmd because support for it is disabled.
Description of all currently known issues is added to README file.

Change-Id: I7f33b79bd2400b4a79adff2673bf83a840ed8ca1
Signed-off-by: Szymon Sliwa <szs@semihalf.com>
Signed-off-by: Michal Mazur <mkm@semihalf.com>